+++ frontend/src/pages/loginpage.jsx
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from "react";
 import "../index.css";
 import NewAccountPage from "./createpage";
-import axiosClient from "../axios.js";
 import { useStateContext } from "../context/ContextProvider.jsx";
 import CountdownTimer from "../context/CountdownTimer.jsx";
+import login from "../utils/login.js";
+import LoginForm from "../../components/LoginForm.jsx";
 
 const LoginPage = ({ onLogin, onLogout, onCreateNewUser }) => {
   const [email, setEmail] = useState("");
@@ -29,39 +30,7 @@ const LoginPage = ({ onLogin, onLogout, onCreateNewUser }) => {
   // Handles the login function
   const handleLogin = (event) => {
     event.preventDefault();
-
-    const credentials = {
-      email: email,
-      password: password,
-    };
-
-    // Checks if the credentials match using tokens, and if the user is authenticated it saves the logged user to local storage
-    axiosClient
-      .post("/token/", credentials)
-      .then(({ data }) => {
-        setToken(data.access);
-        axiosClient
-          .get("/users/userinfo", {
-            headers: {
-              Authorization: `Bearer ${data.access}`,
-            },
-          })
-          .then((response) => {
-            setUser(response.data);
-            console.log("User details:", response.data);
-            localStorage.setItem("loggedUser", JSON.stringify(response.data));
-            localStorage.setItem("isLoggedIn", true);
-            onLogin();
-          });
-      })
-      .catch((err) => {
-        const response = err.response;
-        if (response && response.status === 422) {
-          setError(response.data.message);
-        } else {
-          setError("Sähköposti tai salasana virheellinen!");
-        }
-      });
+    login({ email, password, setError, setToken, onLogin, setUser });
   };
 
   // Handles the logout function and clears the countdown timer
@@ -93,38 +62,15 @@ const LoginPage = ({ onLogin, onLogout, onCreateNewUser }) => {
           <CountdownTimer initialTime={timeLeft} onExpire={handleLogout} />
         </>
       ) : (
-        <form>
-          <h3>Sisäänkirjautuminen</h3>
-          <div className="form-group">
-            <label htmlFor="email">Sähköposti:</label>
-            <input
-              id="email"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="password">Salasana:</label>
-            <input
-              id="password"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-          </div>
-          <button className="login-button" type="submit" onClick={handleLogin}>
-            Kirjaudu sisään
-          </button>
-          <button
-            className="create-user-button"
-            type="button"
-            onClick={handleCreateUser}
-          >
-            Luo uusi käyttäjä
-          </button>
-          {error && <div style={{ color: "red" }}>{error}</div>}
-        </form>
+        <LoginForm
+          email={email}
+          password={password}
+          setEmail={setEmail}
+          setPassword={setPassword}
+          handleLogin={handleLogin}
+          handleCreateUser={handleCreateUser}
+          error={error}
+        />
       )}
     </div>
   );
